change service classes

+++ src/main/java/com/andrii/footballmanager/repo/PlayerRepository.java
@@ -3,5 +3,5 @@ package com.andrii.footballmanager.repo;
 import com.andrii.footballmanager.entity.Player;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlayerRepository extends CrudRepository<Player, Integer> {
+public interface PlayerRepository extends CrudRepository<Player, Long> {
 }
