+++ src/main/java/com/andrii/footballmanager/repo/TeamRepository.java
@@ -3,5 +3,5 @@ package com.andrii.footballmanager.repo;
 import com.andrii.footballmanager.entity.Team;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TeamRepository extends CrudRepository<Team, Integer> {
+public interface TeamRepository extends CrudRepository<Team, Long> {
 }
